Add buildvfs, abstraction layer for file I/O.
Currently it passes calls through to the system libraries as before.

Also adds an incomplete implementation on PhysFS.

git-svn-id: https://svn.eduke32.com/eduke32@7359 1a8010ca-5511-0410-912e-c29ae57300e0

+++ source/mact/src/animlib.cpp
@@ -121,7 +121,7 @@ static inline uint16_t findpage(uint16_t framenumber)
         {
             lp_descriptor & lp = anim->LpArray[i];
             if (lp.baseRecord <= framenumber && framenumber < lp.baseRecord + lp.nRecords)
-                return i;
+                return (uint16_t)i;
         }
 
         if (j && i == nLps)
@@ -134,7 +134,7 @@ static inline uint16_t findpage(uint16_t framenumber)
     }
     while (0);
 
-    return i;
+    return (uint16_t)i;
 }
